extract out mutation hook + error state
